code/requirements.txt
numpy==1.26.4
pandas==2.2.1
PyYAML==6.0.1
scipy==1.12.0